+++ src/coreMethods/cardToken/type.ts
@@ -1,0 +1,6 @@
+export type TFieldsCardTokenParams = {
+  cardId?: string;
+  cardholderName?: string;
+  identificationType?: string;
+  identificationNumber?: string;
+};
