+++ src/mercadoPago/initMercadoPago/type.ts
@@ -1,8 +1,10 @@
-import type { TCardToken, TCardTokenParams } from "../../coreMethods/createCardToken/types";
+import type { TFieldsCardTokenParams } from "../../coreMethods/cardToken/type";
+import type { TCardTokenParams } from "../../coreMethods/createCardToken/types";
 import type { IdentificationType } from "../../coreMethods/getIdentificationTypes/types";
 import type { TInstallments, TInstallmentsParams } from "../../coreMethods/getInstallments/types";
 import type { TIssuers, TIssuersParams } from "../../coreMethods/getIssuers/types";
 import type { TPaymentMethods, TPaymentMethodsParams } from "../../coreMethods/getPaymentMethods/types";
+import type { TCardToken } from "../../coreMethods/util/types";
 
 export type TOptions = {
   /**
@@ -33,4 +35,7 @@ export type TInstanceMercadoPago = {
   getIssuers: (issuersParams: TIssuersParams) => Promise<TIssuers[]>;
   getInstallments: (installmentsParams: TInstallmentsParams) => Promise<TInstallments[]>;
   createCardToken: (cardTokenParams: TCardTokenParams) => Promise<TCardToken>;
+  fields: {
+    createCardToken: (fieldsCardTokenParams: TFieldsCardTokenParams) => Promise<TCardToken>;
+  };
 };
